refactor(lints): move lints definitions to a rules module
