+++ src/cargo/lints/rules/im_a_teapot.rs
