+++ src/cargo/lints/rules/mod.rs
@@ -1,0 +1,16 @@
+mod blanket_hint_mostly_unused;
+mod im_a_teapot;
+mod implicit_minimum_version_req;
+mod unknown_lints;
+
+pub use blanket_hint_mostly_unused::blanket_hint_mostly_unused;
+pub use im_a_teapot::check_im_a_teapot;
+pub use implicit_minimum_version_req::implicit_minimum_version_req;
+pub use unknown_lints::output_unknown_lints;
+
+pub const LINTS: &[crate::lints::Lint] = &[
+    blanket_hint_mostly_unused::LINT,
+    implicit_minimum_version_req::LINT,
+    im_a_teapot::LINT,
+    unknown_lints::LINT,
+];
